+++ components/RegisterNotice/RegisterNotice.types.ts
