+++ pages/test.tsx
@@ -1,4 +1,4 @@
-import Register from "@/components/RegisterNotice/test";
+import Register from "@/components/RegisterNotice/index";
 
 export default function test() {
   return <Register />;
